shared/format_standardization/format_converter.py
"""
Format Converter
Converts between different model and visualization formats
"""

from pathlib import Path
from typing import Dict, Any, Optional, List
import logging
from .gltf_processor import GLTFProcessor
from .svg_processor import SVGProcessor

logger = logging.getLogger(__name__)


class FormatConverter:
    """
    Format converter for model and visualization assets
    Supports conversions between GLTF, SVG, and other formats
    """

    def __init__(self):
        self.gltf_processor = GLTFProcessor()
        self.svg_processor = SVGProcessor()
        self.conversion_matrix = self._build_conversion_matrix()

    def _build_conversion_matrix(self) -> Dict[str, List[str]]:
        """Build matrix of supported format conversions"""
        return {
            'gltf': ['gltf', 'glb', 'obj'],  # GLTF can be converted to other 3D formats
            'glb': ['gltf', 'obj'],  # GLB (binary GLTF) conversions
            'svg': ['svg', 'png', 'jpg'],  # SVG can be rasterized
            'obj': ['gltf', 'glb'],  # OBJ to GLTF conversions
            'png': ['svg'],  # Limited PNG to vector conversions
            'jpg': ['svg']   # Limited JPG to vector conversions
        }

    def can_convert(self, from_format: str, to_format: str) -> bool:
        """Check if conversion between formats is supported"""
        from_format = from_format.lower()
        to_format = to_format.lower()

        if from_format == to_format:
            return True  # No conversion needed

        return to_format in self.conversion_matrix.get(from_format, [])

    def convert_file(self, input_path: Path, output_path: Path,
                    target_format: Optional[str] = None) -> Optional[Path]:
        """Convert a file to the target format"""
        if not input_path.exists():
            logger.error(f"Input file does not exist: {input_path}")
            return None

        # Determine target format from output path if not specified
        if not target_format:
            target_format = output_path.suffix[1:].lower()  # Remove the dot

        input_format = input_path.suffix[1:].lower()  # Remove the dot

        # Check if conversion is supported
        if not self.can_convert(input_format, target_format):
            logger.error(f"Conversion from {input_format} to {target_format} is not supported")
            return None

        # Perform conversion
        try:
            if input_format in ['gltf', 'glb'] and target_format in ['gltf', 'glb']:
                return self._convert_gltf_formats(input_path, output_path, target_format)
            elif input_format == 'svg' and target_format in ['png', 'jpg']:
                return self._convert_svg_to_raster(input_path, output_path, target_format)
            elif input_format in ['png', 'jpg'] and target_format == 'svg':
                return self._convert_raster_to_svg(input_path, output_path)
            elif input_format == 'obj' and target_format in ['gltf', 'glb']:
                return self._convert_obj_to_gltf(input_path, output_path, target_format)
            else:
                # Same format or unsupported conversion
                logger.warning(f"No conversion needed or supported for {input_format} -> {target_format}")
                return None

        except Exception as e:
            logger.error(f"Conversion failed: {e}")
            return None

    def _convert_gltf_formats(self, input_path: Path, output_path: Path, target_format: str) -> Optional[Path]:
        """Convert between GLTF formats (GLTF <-> GLB)"""
        # This would implement actual GLTF/GLB conversion
        # For now, just copy the file
        import shutil
        shutil.copy2(input_path, output_path)
        logger.info(f"Copied GLTF file (conversion not yet implemented): {output_path}")
        return output_path

    def _convert_svg_to_raster(self, input_path: Path, output_path: Path, target_format: str) -> Optional[Path]:
        """Convert SVG to raster format (PNG/JPG)"""
        # This would use a library like CairoSVG or Inkscape to rasterize
        # For now, create a placeholder
        try:
            # Placeholder: create a simple colored square image
            # In real implementation, would use proper SVG rasterization
            if target_format in ['png', 'jpg']:
                # Create a minimal valid image file (this is just a placeholder)
                with open(output_path, 'wb') as f:
                    # Write minimal PNG header (this won't display properly)
                    f.write(b'\x89PNG\r\n\x1a\n')  # PNG signature
                logger.warning(f"SVG to {target_format} conversion not fully implemented: {output_path}")
                return output_path
        except Exception as e:
            logger.error(f"SVG rasterization failed: {e}")
            return None

    def _convert_raster_to_svg(self, input_path: Path, output_path: Path) -> Optional[Path]:
        """Convert raster image to SVG (limited support)"""
        # This is very limited - would need advanced image tracing
        # For now, create a placeholder SVG
        try:
            svg_content = '''<?xml version="1.0" encoding="UTF-8"?>
<svg width="64" height="64" xmlns="http://www.w3.org/2000/svg">
  <rect width="64" height="64" fill="#cccccc" stroke="#999999" stroke-width="1"/>
  <text x="32" y="35" text-anchor="middle" font-family="Arial" font-size="12" fill="#666666">Image</text>
</svg>'''

            with open(output_path, 'w') as f:
                f.write(svg_content)

            logger.warning(f"Created placeholder SVG (tracing not implemented): {output_path}")
            return output_path

        except Exception as e:
            logger.error(f"Raster to SVG conversion failed: {e}")
            return None

    def _convert_obj_to_gltf(self, input_path: Path, output_path: Path, target_format: str) -> Optional[Path]:
        """Convert OBJ to GLTF format"""
        # This would use a library like pygltflib or assimp
        # For now, create a placeholder GLTF
        try:
            gltf_data = {
                "asset": {
                    "version": "2.0",
                    "generator": "Integrated Network Platform Converter"
                },
                "scene": 0,
                "scenes": [{"name": "OBJ Import"}],
                "nodes": [{
                    "name": "ImportedOBJ",
                    "mesh": 0
                }],
                "meshes": [{
                    "name": "OBJMesh",
                    "primitives": [{
                        "mode": 4,  # TRIANGLES
                        "attributes": {
                            "POSITION": 0
                        }
                    }]
                }],
                "buffers": [{
                    "uri": "data:application/octet-stream;base64,AAABAAIAAAAAAAAAAAAAAAAAAAAAAIA/AAAAAAAAAAAAAAAAAACAPwAAAAA=",
                    "byteLength": 44
                }],
                "bufferViews": [{
                    "buffer": 0,
                    "byteOffset": 0,
                    "byteLength": 44
                }],
                "accessors": [{
                    "bufferView": 0,
                    "componentType": 5126,  # FLOAT
                    "count": 3,
                    "type": "VEC3",
                    "max": [1.0, 1.0, 1.0],
                    "min": [-1.0, -1.0, -1.0]
                }]
            }

            import json
            with open(output_path, 'w') as f:
                json.dump(gltf_data, f, indent=2)

            logger.warning(f"Created placeholder GLTF (OBJ import not implemented): {output_path}")
            return output_path

        except Exception as e:
            logger.error(f"OBJ to GLTF conversion failed: {e}")
            return None

    def batch_convert(self, input_directory: Path, output_directory: Path,
                     target_format: str, source_formats: Optional[List[str]] = None) -> Dict[str, Any]:
        """Batch convert files in a directory"""
        results = {
            'converted': 0,
            'failed': 0,
            'skipped': 0,
            'files': []
        }

        if not source_formats:
            source_formats = list(self.conversion_matrix.keys())

        # Create output directory
        output_directory.mkdir(parents=True, exist_ok=True)

        # Find files to convert
        files_to_convert = []
        for fmt in source_formats:
            files_to_convert.extend(list(input_directory.glob(f"*.{fmt}")))

        results['total'] = len(files_to_convert)

        for input_file in files_to_convert:
            output_file = output_directory / f"{input_file.stem}.{target_format}"

            if output_file.exists():
                results['skipped'] += 1
                results['files'].append({
                    'input': str(input_file),
                    'output': str(output_file),
                    'status': 'skipped',
                    'reason': 'Output file already exists'
                })
                continue

            converted_path = self.convert_file(input_file, output_file, target_format)

            if converted_path:
                results['converted'] += 1
                results['files'].append({
                    'input': str(input_file),
                    'output': str(converted_path),
                    'status': 'converted'
                })
            else:
                results['failed'] += 1
                results['files'].append({
                    'input': str(input_file),
                    'output': str(output_file),
                    'status': 'failed'
                })

        logger.info(f"Batch conversion complete: {results['converted']} converted, {results['failed']} failed, {results['skipped']} skipped")
        return results

    def get_supported_conversions(self) -> Dict[str, List[str]]:
        """Get all supported format conversions"""
        return self.conversion_matrix.copy()

    def validate_conversion_requirements(self) -> Dict[str, Any]:
        """Check if all conversion requirements are met"""
        requirements = {
            'gltf_processing': True,  # Built-in
            'svg_processing': True,   # Built-in
            'raster_conversion': False,  # Would need external libraries
            'obj_import': False,      # Would need external libraries
            'external_tools': []
        }

        # Check for optional external tools/libraries
        try:
            import PIL
            requirements['raster_conversion'] = True
        except ImportError:
            requirements['external_tools'].append('Pillow (for raster image processing)')

        try:
            import pygltflib
            requirements['obj_import'] = True
        except ImportError:
            requirements['external_tools'].append('pygltflib (for 3D model conversion)')

        return requirements